Re-implemented var.combine_meta()  - was removed when a changeset was unrolled.
Started to remove superfluous references to 'atts' and 'plotatts' in the code.
The transfer of metadata from one variable to another should be delegated to
either combine_meta() or copy_meta(), not done manually in each operator.

Did you know there are 103 references to 'plotatts' outside the 'plot' module,
in 20 different files?  All to make the output of one method (plot.plotvar)
look prettier!  This attribute is like a shit stain that's streaked across the
underwear of our codebase.  It needs some serious refactoring!