Implemented timeutils.jointimeaxes().
Added _splittime, _jointime helper functions to timeutils to handle the
splitting/joining for time axes themselves.  This is used heavily in the
corresponding split/join Var operators, since we need to switch back and
forth between each representation when determining what input data we need
for the given output slice.