Be more careful when doing arithmetic on an array in reduce.py
We may be dealing with a degenerate array (0-dimensional), in which case any
arithmetic operations will automatically transform the result to a scalar.

If this scalar is then accessed as an array, PyGeode will crash.

Fixes issue 32.

Note:  I noticed other non-critical places where the automatic scalarization
is happening (such as in nanvariance).  However, I think that particular case
would just trigger a warning in View.get and carry on.  Since it's not part of
a bug report, I'm leaving it for now.  That change should go directly in the
development branch, not the release branch.

--HG--
branch : release-0.6