In ufunc, keep the original variable name if we're doing an arithmetic scalar operation on the variable.
This is a common case (e.g. for unit conversion), in which the original name
is probably more useful than something mangled with parentheses and arithmetic
symbols.

If you want the old behaviour (mangling the name into an arithmetic expression),
you can try adding parentheses to the name of the var you're starting with.