Change the behaviour of Var.__getitem__, so that degenerate axes are removed.
This should be similar to what happens when slicing raw numpy arrays.

E.g., x[0,:,:] would select the first index of the leftmost row, then remove
that dimension.  The result would be a 2-dimensional array.

Note: This changes the behaviour of the [] operator in PyGeode, so you should
check that your old scripts still work as expected!

Fixes issue 49.