Modify the test for issue 41.
The two sets of latitudes should NOT be mappable, so make sure we crash right
when we try to do the broadcast.  If we didn't crash, then we're left with an
inconsistent object that will crash deep in the view stack when we try to
retrieve the values!

--HG--
branch : release-1.0