Forgot parentheses around an "if" condition in interp.c.
By luck, isfinite() was a macro which returned something enclosed in
parentheses, so the compiler was happy.

However, this macro seems to have changed in recent versions of libc, so now
we're getting a syntax error on that line.

Uncovered while trying to build a package for Ubuntu xenial from launchpad.