Fix bug in tools.map_to when len(a) == 0.
This was not properly handled in the C code, resulting in an improper memory
access.  Surprisingly, this was never caught in the unit tests when I ran them.
I'm guessing map_to found some garbage value there, which didn't match the b
array, and set '-1' for that index (i.e. no-match).

However, when I re-ran the unit tests with a more recent numpy package
some of them were failing (the concat ones with size 0 axes). I'm guessing
there was a change to the memory layout of numpy or something, and this was
causing whatever 'a' was pointing to to match the 'b' value, giving
a false match.

The annoying thing was, the unit tests were passing when I ran from an in-place
build with stock numpy in Ubuntu 14.04.  However, if I install those exact
same build files into a virtualenv alongside a more recent numpy package,
suddenly the tests failed.  Was scratching my head for a while, wondering if
there was a bug in numpy.