Drop the use of pyproject.toml.
The intention of this file was to specify numpy as a build-time dependency, so
that we're guaranteed the user has numpy installed before building from source.
This was supposed to eliminate build failures when numpy wasn't installed.

However, the existence of this file also seems to trigger an "isolated build
environment", which installs the pyproject.toml dependencies in a separate
build environment before constructing the wheel.  However, this will cause it
to download and install the latest version of numpy for building the extension
modules (without touching the numpy version in the user's virtual environment). After the build, it will revert back to the user's installed version of numpy.

If the latest version of numpy is not ABI compatible with the version that the
user has, this will cause very annoying run-time failures when trying to
import pygeode.